fix: B4 was double-counting the distillation term (ticket 08 review)
The signal returned w*KL while the backbone had already contributed
mean(KL), so B4 optimised mean(KL) + mean(w*KL): twice B2's distillation
weight, and the paper's 4:1 high-to-low token ratio flattened to 2.15:1.
Both versions train and both report a plausible number, which is the
failure mode this repo exists to avoid. The signal now returns the
residual (w - 1)*KL, so the step loss is exactly VA-OPD's L_group, B4's
backbone configuration stays identical to B2's, and B4 reduces to B2
exactly on an example whose weights come back uniform.

TrainingConfig.coupling_threshold, added for a gate no baseline uses,
changed the run hash of every trained Group — including the B2 hash
published beside a merged checkpoint. Removed; the gate keeps its own
default. All three baseline hashes are now pinned by a test, so the next
schema field cannot quietly retrain them.

Also from review: the store schema had two definitions and they had
already diverged, so both paths now derive from store_schema; the router
capture keyed by depth rather than by the layer's own index, which would
read the wrong layer without raising on a model whose layers are not all
hybrid; gold_logprobs annotated an id tensor as Float, so beartype could
never have caught a rank error there; build_signals carried an if/elif
cascade duplicating the registry beside it; LayerMapping.kind becomes a
StrEnum like the other closed sets; and a parameter used "head", which
CONTEXT.md lists as a term to avoid for a probe.

+++ src/ceed_teacher/ceed_teacher/va_opd.py
@@ -29,7 +29,7 @@ from __future__ import annotations
 from typing import Any
 
 import torch
-from jaxtyping import Float
+from jaxtyping import Float, Int
 from torch import Tensor
 
 VISUAL_ADVANTAGE = "visual_advantage"
@@ -99,7 +99,7 @@ def visual_advantage(
 
 
 def gold_logprobs(
-    logits: Float[Tensor, "positions vocab"], gold_token_ids: Float[Tensor, " tokens"]
+    logits: Float[Tensor, "positions vocab"], gold_token_ids: Int[Tensor, " tokens"]
 ) -> Float[Tensor, " tokens"]:
     """Return the teacher's log-probability of each gold token at its position.
